Support non Sync + Send error types via shim. Move InsertError into error module and simplify with thiserror.
